# faster ftrl fit and predict based on many c-style optimization