I worked on some cython optimization on `fit` and `predict` function. The speed is 2X~5X when `interaction=False`:

### before
```
       3 function calls in 4.597 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.000    0.000 base.py:99(get_shape)
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
        1    4.597    4.597    4.597    4.597 {method 'fit' of 'kaggler.online_model.ftrl.FTRL' objects}

```

### after
```
         3 function calls in 0.981 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.000    0.000 base.py:99(get_shape)
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
        1    0.981    0.981    0.981    0.981 {method 'fit' of 'kaggler.online_model.ftrl.FTRL' objects}
```

Note that when `interaction=True`, the main overhead will be:
```python
abs(hash('{}_{}'.format(x[i], x[j])))
```
I've provide the solution for it and commented it out in the `_indices` function. I've try that out and find it really solve the overhead. Plz take a look and see if it works general enough.